36 - small testing reorganization
