+++ chimp.config.js
@@ -1,3 +1,5 @@
 module.exports = {
-  path: './test/functional-tests'
+  path: './test/functional-tests',
+  mocha: true,
+  chai: true
 }
