+++ test/functional-tests/.eslintrc
@@ -1,0 +1,8 @@
+{
+  "env": {
+    "mocha": true,
+  },
+  "globals": {
+    "browser":
+  }
+}
