+++ test/functional-tests/player.test.js
@@ -44,7 +44,6 @@ describe('Player: ', function () {
 
   before(async function () {
 
-    await paratii.eth.deployContracts()
     await paratii.core.vids.create({
       id: 'foo',
       owner: address1,
