+++ test/functional-tests/test-utils/helpers.js
@@ -8,15 +8,7 @@
 // */
 //
 
-/* global localStorage */
 import { Paratii, utils } from 'paratii-lib'
-// import { XMLHttpRequest } from 'xmlhttprequest'
-import { assert } from 'chai'
-
-var chai = require('chai')
-var chaiAsPromised = require('chai-as-promised')
-
-chai.use(chaiAsPromised)
 
 // this address will be used as the owner address for all paratii contracts in the tests
 let address = '0x9e2d04eef5b16CFfB4328Ddd027B55736407B275'
@@ -50,116 +42,6 @@ export const paratii = new Paratii({
 // const add0x = utils.add0x
 //
 
-// // The before  function will be run once, before all tests
-before(async function (done) {
-  browser.addCommand('waitForClickable', function (selector, timeout) {
-    this.waitForVisible(selector, timeout)
-    this.waitForEnabled(selector, timeout)
-  })
-
-  browser.addCommand('waitUntilVideoIsPlaying', () => {
-    browser.waitUntil(() => (
-      parseInt(browser.getAttribute('#video-player', 'currentTime'), 10) !== 0 &&
-			browser.getAttribute('#video-player', 'paused') !== 'true' &&
-			browser.getAttribute('#video-player', 'ended') !== 'true'
-    ))
-  })
-
-  browser.addCommand('waitUntilBuffered', () => {
-    browser.waitUntil(() => !!browser.execute(() => {
-      const playerEl = document.querySelector('#video-player')
-      for (let i = 0; i < playerEl.buffered.length; i += 1) {
-        if (playerEl.buffered.end(i) > 0) {
-          return true
-        }
-      }
-      return false
-    }).value)
-  })
-  //
-  //   browser.addCommand('waitAndSetValue', function (selector, value, timeout) {
-  //     this.waitForVisible(selector, timeout)
-  //     this.waitForEnabled(selector, timeout)
-  //     browser.waitUntil(function () {
-  //       try {
-  //         browser.setValue(selector, value)
-  //         return true
-  //       } catch (err) {
-  //         if (err.seleniumStack.type === 'InvalidElementState') {
-  //           // ignore and try again
-  //           return false
-  //         } else {
-  //           throw err
-  //         }
-  //       }
-  //     }, timeout, `Could not set value on ${selector} (timeout: ${timeout}s)`)
-  //   })
-  browser.addCommand('waitAndClick', function (selector, timeout) {
-    this.waitForVisible(selector, timeout)
-    this.waitForEnabled(selector, timeout)
-    browser.waitUntil(function () {
-      try {
-        browser.click(selector)
-        return true
-      } catch (err) {
-        if (err.seleniumStack.type === 'InvalidElementState') {
-          // ignore and try again
-          return false
-        } else if (err.seleniumStack.type === 'UnknownError') {
-          // 'another element would receive the click' is reported as an 'unknown error'
-          // ignore and try again
-          return false
-        } else {
-          console.log(err)
-          throw err
-        }
-      }
-    }, timeout, `Could not click on ${selector} (timeout: ${timeout}s)`)
-  })
-  //   browser.addCommand('waitUntilRequestHasStatus', function (url, status = 200, method = 'GET', timeout) {
-  //     browser.waitUntil(() => {
-  //       const request = new XMLHttpRequest()
-  //       request.open(method, url, false)
-  //       request.send(null)
-  //       return request.status === status
-  //     }, timeout, `The ${method} request to ${url} never achieved a ${status} status`)
-  //   })
-  //
-  browser.addCommand('waitAndRemove', function (selector, timeout) {
-    this.waitForVisible(selector)
-    browser.execute((selectorToRemove) => {
-      const element = document.querySelector(selectorToRemove)
-      if (element) {
-        element.remove()
-      }
-    }, selector)
-  })
-  //
-  //   browser.addCommand('sendSomeETH', async function (beneficiary, amount, timeout) {
-  //     // console.log(`send ${amount} to ${beneficiary}`)
-  //     await sendSomeETH(beneficiary, amount)
-  //     browser.waitUntil(function () {
-  //       let result = browser.execute(function () {
-  //         return Session.get('eth_balance')
-  //       })
-  //       return result.value && result.value > 0
-  //     }, timeout, `the ETH did not arrive..`)
-  //   })
-  //
-  //   browser.addCommand('sendSomePTI', async function (beneficiary, amount, timeout) {
-  //     await sendSomePTI(beneficiary, amount)
-  //     browser.waitUntil(function () {
-  //       let result = browser.execute(function () {
-  //         return Session.get('pti_balance')
-  //       })
-  //       // console.log(`PTI balance: ${result.value}`)
-  //       return result.value && result.value > 0
-  //     }, timeout, `the PTI did not arrive..`)
-  //   })
-  //   browser.url('http://localhost:3000')
-  //   browser.contracts = await getOrDeployParatiiContracts(server, browser)
-  done()
-})
 //
 // // The beforeEach function is run before each single test
 // beforeEach(function () {
