+++ test/unit-tests/.eslintrc
@@ -1,0 +1,3 @@
+"env": {
+  "jasmine": true
+}
